shift to one file
